claude-gateway/requirements.txt
# Core
openai>=1.40.0              # OpenAI-compatible client (works with NVIDIA NIM, Kimi, Groq, Ollama)
python-dotenv>=1.0.0
aiohttp>=3.9.0
aiofiles>=23.2.1
httpx>=0.27.0

# Telegram
python-telegram-bot>=21.0

# Discord
discord.py>=2.3.0

# WhatsApp (requires Go 1.21+ on the system: sudo apt install golang-go)
neonize>=0.2.0

# iMessage (BlueBubbles) — uses websockets + aiohttp above
websockets>=12.0

# Slack (Socket Mode — optional)
slack-bolt>=1.18.0

# Media handling
Pillow>=10.0.0
